feat inclui user Venda

+++ frontend/src/app/services/vendas.service.ts
@@ -2,14 +2,20 @@ import { httpClient } from "app/http";
 import { Page } from "app/models/common/page";
 import { ItemProdutoAtualizarDTO, Venda } from "app/models/vendas";
 import { AxiosResponse } from "axios";
+import { useUser } from "context/UserContext"; // Importa o contexto do usuário
 
 const resourceURL = "/api/vendas";
 
 export const useVendaService = () => {
+  const { user } = useUser(); // Acessa o usuário logado do contexto
+
   const realizarVenda = async (venda: Venda): Promise<void> => {
-    console.log("Venda enviada para o backend:", venda); 
-    
-    await httpClient.post<Venda>(resourceURL, venda);
+    if (user) {
+      console.log("Venda enviada para o backend:", venda); 
+      await httpClient.post<Venda>(`${resourceURL}?userId=${user.id}`, venda); // Passa o userId
+    } else {
+      throw new Error("Usuário não autenticado");
+    }
   };
 
   const find = async (
@@ -24,56 +30,59 @@ export const useVendaService = () => {
     dataEntregaInicio?: string, 
     dataEntregaFim?: string
   ): Promise<Page<Venda>> => {
+    if (user) {
+      // Montando a URL com os filtros opcionais e o userId
+      let url = `${resourceURL}?nomeCliente=${nomeCliente}&page=${page}&size=${rows}&userId=${user.id}`;
+      
+      if (formaPagamento) url += `&formaPagamento=${formaPagamento}`;
+      if (statusPagamento) url += `&statusPagamento=${statusPagamento}`;
+      if (statusPedido) url += `&statusPedido=${statusPedido}`;
+      if (dataCadastroInicio) url += `&dataCadastroInicio=${dataCadastroInicio}`;
+      if (dataCadastroFim) url += `&dataCadastroFim=${dataCadastroFim}`;
+      if (dataEntregaInicio) url += `&dataEntregaInicio=${dataEntregaInicio}`;
+      if (dataEntregaFim) url += `&dataEntregaFim=${dataEntregaFim}`;
     
-    // Montando a URL com os filtros opcionais
-    let url = `${resourceURL}?nomeCliente=${nomeCliente}&page=${page}&size=${rows}`;
-    
-    // Adiciona os parâmetros de filtro à URL se estiverem definidos
-    if (formaPagamento) {
-      url += `&formaPagamento=${formaPagamento}`;
+      // Faz a requisição e retorna os dados
+      const response: AxiosResponse<Page<Venda>> = await httpClient.get(url);
+      return response.data;
+    } else {
+      throw new Error("Usuário não autenticado");
     }
-    if (statusPagamento) {
-      url += `&statusPagamento=${statusPagamento}`;
-    }
-    if (statusPedido) {
-      url += `&statusPedido=${statusPedido}`;
-    }
-    if (dataCadastroInicio) {
-      url += `&dataCadastroInicio=${dataCadastroInicio}`;
-    }
-    if (dataCadastroFim) {
-      url += `&dataCadastroFim=${dataCadastroFim}`;
-    }
-    if (dataEntregaInicio) {
-      url += `&dataEntregaInicio=${dataEntregaInicio}`;
-    }
-    if (dataEntregaFim) {
-      url += `&dataEntregaFim=${dataEntregaFim}`;
-    }
-  
-    // Faz a requisição e retorna os dados
-    const response: AxiosResponse<Page<Venda>> = await httpClient.get(url);
-    return response.data;
   };
   
-    // Nova função para buscar a venda por ID
-    const buscarPorId = async (id: number): Promise<Venda> => {
-      const response: AxiosResponse<Venda> = await httpClient.get(`${resourceURL}/${id}`);
+  const buscarPorId = async (id: number): Promise<Venda> => {
+    if (user) {
+      const response: AxiosResponse<Venda> = await httpClient.get(`${resourceURL}/${id}?userId=${user.id}`);
       return response.data;
-    };
+    } else {
+      throw new Error("Usuário não autenticado");
+    }
+  };
 
   const deletar = async (id: number): Promise<AxiosResponse<void>> => {
-    return await httpClient.delete(`${resourceURL}/${id}`);
+    if (user) {
+      return await httpClient.delete(`${resourceURL}/${id}?userId=${user.id}`);
+    } else {
+      throw new Error("Usuário não autenticado");
+    }
   };
 
-    const listarItensPorVenda = async (idVenda: number): Promise<ItemProdutoAtualizarDTO[]> => {
-      const response: AxiosResponse<ItemProdutoAtualizarDTO[]> = await httpClient.get(`${resourceURL}/${idVenda}/itens-produto`);
+  const listarItensPorVenda = async (idVenda: number): Promise<ItemProdutoAtualizarDTO[]> => {
+    if (user) {
+      const response: AxiosResponse<ItemProdutoAtualizarDTO[]> = await httpClient.get(`${resourceURL}/${idVenda}/itens-produto?userId=${user.id}`);
       return response.data;
-    };
+    } else {
+      throw new Error("Usuário não autenticado");
+    }
+  };
 
-    const atualizarVenda = async (id: number, vendaAtualizadaDTO: Venda): Promise<void> => {
-      await httpClient.put<Venda>(`${resourceURL}/${id}`, vendaAtualizadaDTO);
-    };
+  const atualizarVenda = async (id: number, vendaAtualizadaDTO: Venda): Promise<void> => {
+    if (user) {
+      await httpClient.put<Venda>(`${resourceURL}/${id}?userId=${user.id}`, vendaAtualizadaDTO);
+    } else {
+      throw new Error("Usuário não autenticado");
+    }
+  };
 
   return {
     realizarVenda,
@@ -83,4 +92,4 @@ export const useVendaService = () => {
     listarItensPorVenda,
     atualizarVenda, 
   };
-}
+};
